fix(backend_test): accelerated readout_correction_retries test
Set a new batch size to  be 2 rather than 300 and changed the max number of jobs to just be 2.
This way we will still test that batches are excuted properly as well as testing handling multiple jobs at once.
The test still runs between 10 - 90 seconds, but this might be as good as it gets for what we are trying to test.